Merge Staci Branch - All Merge Conflicts Resolved.

+++ src/Model/AcidSpitter.java
@@ -1,9 +1,10 @@
 package Model;
 
-public class AcidSpitter extends Glyphid {
+public class
+AcidSpitter extends Glyphid {
     public AcidSpitter(String theName, double theX, double theY, int theHealth,
-                       double theWidth, double theHeight, double theMoveSpeed,
-                       Weapon theWeapon, int fireTimer) {
+                        int theWidth, int theHeight, double theMoveSpeed,
+                        Weapon theWeapon, int fireTimer) {
         super(theName, theX, theY, theHealth, theWidth, theHeight, theMoveSpeed, theWeapon, fireTimer);
     }
