+++ src/Model/Glyphid.java
@@ -4,7 +4,7 @@ public abstract class Glyphid extends Character {
     protected int myFireTimer;
 
     public Glyphid(String theName, double theX, double theY, int theHealth,
-                   double theWidth, double theHeight, double theMoveSpeed,
+                   int theWidth, int theHeight, double theMoveSpeed,
                    Weapon theWeapon, int fireTimer) {
         super(theName, theX, theY, theHealth, theWidth, theHeight, theMoveSpeed, theWeapon);
         this.myFireTimer = fireTimer;
