+++ src/Model/Praetorian.java
@@ -1,7 +1,7 @@
 package Model;
 
 public class Praetorian extends Grunt{
-    public Praetorian(String theName, double theX, double theY, int theHealth, double theWidth, double theHeight, double theMoveSpeed, Weapon theWeapon, int fireTimer) {
+    public Praetorian(String theName, double theX, double theY, int theHealth, int theWidth, int theHeight, double theMoveSpeed, Weapon theWeapon, int fireTimer) {
         super(theName, theX, theY, theHealth, theWidth, theHeight, theMoveSpeed, theWeapon, fireTimer);
     }
 }
