+++ src/Model/Rock.java
@@ -2,7 +2,7 @@ package Model;
 
 public class Rock extends Glyphid {
     public Rock(String theName, double theX, double theY, int theHealth,
-                double theWidth, double theHeight, double theMoveSpeed,
+                int theWidth, int theHeight, double theMoveSpeed,
                 Weapon theWeapon, int fireTimer) {
         super(theName, theX, theY, theHealth, theWidth, theHeight, theMoveSpeed, theWeapon, fireTimer);
     }
